backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-jose[cryptography]
sqlalchemy
psycopg[binary]
python-dotenv
